Refactoring to allow for configurable Steam directories
Split out the Steam directory selection into a separate data structure
which gets loaded from a file. Presently, there isn't a way to create
the file, but once a UI is built and it writes to this file, it will
allow the user to configure the paths for both the rimworld save
directory and the app installation itself.